+++ scripts/upgrade.py
@@ -72,10 +72,11 @@ from migrations.v120_to_v121 import Migration120to121
 from migrations.v121_to_v130 import Migration121to130
 from migrations.v130_to_v140 import Migration130to140
 from migrations.v140_to_v150 import Migration140to150
+from migrations.v150_to_v151 import Migration150to151
 
 
 # Latest version
-LATEST_VERSION = "1.5.0"
+LATEST_VERSION = "1.5.1"
 
 # All available migrations in order
 MIGRATIONS = [
@@ -108,6 +109,7 @@ MIGRATIONS = [
     Migration121to130,
     Migration130to140,
     Migration140to150,
+    Migration150to151,
 ]